Tree scan statistics for R

Tree scan statistics for R

Usage

TreeMineR(
  data,
  tree,
  p = NULL,
  n_exposed = NULL,
  n_unexposed = NULL,
  dictionary = NULL,
  delimiter = "/",
  n_monte_carlo_sim = 9999,
  random_seed = FALSE,
  future_control = list(strategy = "sequential")
)

Arguments

data

The dataset used for the computation. The dataset needs to include the following columns:

`id`

An integer that is unique to every individual.

`leaf`

A string identifying the unique diagnoses or leafs for each individual.

`exposed`

A 0/1 indicator of the individual's exposure status.

See below for the first and last rows included in the example dataset.

``` id leaf exposed 1 K251 0 2 Q702 0 3 G96 0 3 S949 0 4 S951 0 --- 999 V539 1 999 V625 1 999 G823 1 1000 L42 1 1000 T524 1 ```

tree

A dataset with one variable `pathString` defining the tree structure that you would like to use. This dataset can, e.g., be created using create_tree.

p

The proportion of exposed individuals in the dataset. Will be calculated based on `n_exposed`, and `n_unexposed` if both are supplied.

n_exposed

Number of exposed individuals (Optional).

n_unexposed

Number of unexposed individuals (Optional).

dictionary

A data.frame that includes one `node` column and a `title` column, which are used for labeling the cuts in the output of `TreeMineR`.

delimiter

A character defining the delimiter of different tree levels within your `pathString`. The default is `/`.

n_monte_carlo_sim

The number of Monte-Carlo simulations to be used for calculating P-values.

random_seed

Random seed used for the Monte-Carlo simulations.

future_control

A list of arguments passed `future::plan`. This is useful if one would like to parallelise the Monte-Carlo simulations to decrease the computation time. The default is a sequential run of the Monte-Carlo simulations.

References

Kulldorff et al. (2003) A tree-based scan statistic for database disease surveillance. Biometrics 56(2): 323–331. DOI: 10.1111/1541-0420.00039.

Examples

TreeMineR(data = diagnoses,
          tree  = icd_10_se,
          p = 1/11,
          n_monte_carlo_sim = 99,
          random_seed = 1234)
#>           cut    n1    n0      llr     p
#>        <char> <int> <int>    <num> <num>
#>    1:      12   122   669 16.18714  0.01
#>    2:      11   132   782 13.65786  0.01
#>    3: V01-X59   241  1687 12.26816  0.01
#>    4: V01-V99   210  1438 11.95732  0.01
#>    5:      15   133   822 11.79775  0.01
#>   ---                                   
#> 3406:     P94     1    10  0.00000  1.00
#> 3407:     T22     1    22  0.00000  1.00
#> 3408:     Q73     1    11  0.00000  1.00
#> 3409:     D15     1    12  0.00000  1.00
#> 3410:     S99     1    12  0.00000  1.00